wip mpp meta project
