+++ annotations/src/main/kotlin/annotations/higherkind.kt
@@ -1,0 +1,9 @@
+package annotations
+
+import kotlin.annotation.AnnotationRetention.RUNTIME
+import kotlin.annotation.AnnotationTarget.CLASS
+
+@Retention(RUNTIME)
+@Target(CLASS)
+@MustBeDocumented
+annotation class higherkind
