+++ annotations/src/main/kotlin/annotations/processors/HigherKindProccessors.kt
@@ -1,0 +1,14 @@
+package annotations.processors
+
+import com.intellij.lang.annotation.AnnotationHolder
+import com.intellij.lang.annotation.Annotator
+import com.intellij.psi.PsiElement
+
+abstract class HigherKindProcessor : Annotator {
+
+  override fun annotate(element: PsiElement, holder: AnnotationHolder) {
+    sayHello()
+  }
+
+  abstract fun sayHello()
+}
