+++ build.gradle
@@ -1,60 +1,74 @@
-plugins {
-    id 'kotlin-multiplatform' version '1.3.30'
+buildscript {
+  apply from: "$rootDir/gradle/dependencies.gradle"
 }
+
+plugins {
+  id "kotlin-multiplatform" version "1.3.30"
+  id "de.undercouch.download" version "3.4.3"
+}
+
+apply from: "$rootDir/gradle/dependencies.gradle"
+apply from: "gradle/ideaDependencies.gradle"
+
+apply plugin: "maven-publish"
 
 repositories {
-    mavenCentral()
+  mavenCentral()
+  google()
+  gradlePluginPortal()
+  maven {
+    url 'https://jitpack.io'
+  }
+  jcenter()
 }
 
-group 'io.arrow-kt'
-version '0.0.1'
-
-apply plugin: 'maven-publish'
+group "io.arrow-kt"
+version "0.0.1"
 
 kotlin {
-    jvm()
-    js()
-    // For ARM, should be changed to iosArm32 or iosArm64
-    // For Linux, should be changed to e.g. linuxX64
-    // For MacOS, should be changed to e.g. macosX64
-    // For Windows, should be changed to e.g. mingwX64
-    linuxX64("linux")
-    sourceSets {
-        commonMain {
-            dependencies {
-                implementation kotlin('stdlib-common')
-            }
-        }
-        commonTest {
-            dependencies {
-                implementation kotlin('test-common')
-                implementation kotlin('test-annotations-common')
-            }
-        }
-        jvmMain {
-            dependencies {
-                implementation kotlin('stdlib-jdk8')
-            }
-        }
-        jvmTest {
-            dependencies {
-                implementation kotlin('test')
-                implementation kotlin('test-junit')
-            }
-        }
-        jsMain {
-            dependencies {
-                implementation kotlin('stdlib-js')
-            }
-        }
-        jsTest {
-            dependencies {
-                implementation kotlin('test-js')
-            }
-        }
-        linuxMain {
-        }
-        linuxTest {
-        }
+  jvm()
+  js()
+  // For ARM, should be changed to iosArm32 or iosArm64
+  // For Linux, should be changed to e.g. linuxX64
+  // For MacOS, should be changed to e.g. macosX64
+  // For Windows, should be changed to e.g. mingwX64
+  linuxX64("linux")
+  sourceSets {
+    commonMain {
+      dependencies {
+        implementation kotlin('stdlib-common')
+      }
     }
+    commonTest {
+      dependencies {
+        implementation kotlin('test-common')
+        implementation kotlin('test-annotations-common')
+      }
+    }
+    jvmMain {
+      dependencies {
+        implementation kotlin('stdlib-jdk8')
+      }
+    }
+    jvmTest {
+      dependencies {
+        implementation kotlin('test')
+        implementation kotlin('test-junit')
+      }
+    }
+    jsMain {
+      dependencies {
+        implementation kotlin('stdlib-js')
+      }
+    }
+    jsTest {
+      dependencies {
+        implementation kotlin('test-js')
+      }
+    }
+    linuxMain {
+    }
+    linuxTest {
+    }
+  }
 }
