+++ gradle/dependencies.gradle
@@ -1,0 +1,22 @@
+ext.V = [
+    autoValue: "1.6.2",
+    idea: '2018.3.3',
+    junit: "4.12",
+    kotlin: "1.3.30",
+    kotlinpoet: "1.1.0"
+]
+
+ext.deps = [
+    autoValue: "com.google.auto.value:auto-value:${V.autoValue}",
+    autoValueAnnotations: "com.google.auto.value:auto-value-annotations:${V.autoValue}",
+    intellijCore: fileTree(dir: "lib/intellij-core/lib", include: [
+        'idea.jar', 'openapi.jar', 'util.jar', 'trove4j.jar', 'extensions.jar',
+        'picocontainer-1.2.jar', 'asm-all.jar', 'automaton-1.12-1.jar',
+        'platform-api.jar', 'platform-impl.jar'
+    ]),
+    junit: "junit:junit:${V.junit}",
+    kotlin: [
+        reflect: "org.jetbrains.kotlin:kotlin-reflect:${V.kotlin}"
+    ],
+    kotlinPoet: "com.squareup:kotlinpoet:${V.kotlinpoet}"
+]
