+++ gradle/ideaDependencies.gradle
@@ -1,0 +1,20 @@
+apply plugin: 'de.undercouch.download'
+
+task downloadIntellij {
+  def downloadDir = new File(project.rootDir, "lib/download")
+  downloadDir.mkdirs()
+
+  def intellijCoreZip = new File(downloadDir, "intellij-core-${V.idea}.zip")
+
+  download {
+    src "https://www.jetbrains.com/intellij-repository/releases/com/jetbrains/intellij/idea/ideaIC/" +
+            "${V.idea}/ideaIC-${V.idea}.zip"
+    dest intellijCoreZip
+    onlyIfNewer true
+  }
+
+  copy {
+    from zipTree(intellijCoreZip)
+    into new File("lib/intellij-core")
+  }
+}
