+++ src/commonTest/kotlin/meta/SampleTests.kt
@@ -1,0 +1,11 @@
+package meta
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class SampleTests {
+    @Test
+    fun testMe() {
+        assertTrue(Sample().checkMe() > 0)
+    }
+}
