+++ src/jsMain/kotlin/meta/SampleJs.kt
@@ -1,0 +1,9 @@
+package meta
+
+actual class Sample {
+    actual fun checkMe() = 12
+}
+
+actual object Platform {
+    actual val name: String = "JS"
+}
