--- src/jsMain/kotlin/sample/SampleJs.kt
@@ -1,9 +1,0 @@
-package sample
-
-actual class Sample {
-  actual fun checkMe() = 12
-}
-
-actual object Platform {
-  actual val name: String = "JS"
-}
