+++ src/jsTest/kotlin/meta/SampleTestsJS.kt
@@ -1,0 +1,11 @@
+package meta
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class SampleTestsJS {
+    @Test
+    fun testHello() {
+        assertTrue("JS" in hello())
+    }
+}
