--- src/jsTest/kotlin/sample/SampleTestsJS.kt
@@ -1,11 +1,0 @@
-package sample
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-class SampleTestsJS {
-  @Test
-  fun testHello() {
-    assertTrue("JS" in hello())
-  }
-}
