+++ src/jvmMain/kotlin/meta/SampleJvm.kt
@@ -1,0 +1,9 @@
+package meta
+
+actual class Sample {
+    actual fun checkMe() = 42
+}
+
+actual object Platform {
+    actual val name: String = "JVM"
+}
