+++ src/jvmTest/kotlin/meta/SampleTestsJVM.kt
@@ -1,0 +1,11 @@
+package meta
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class SampleTestsJVM {
+    @Test
+    fun testHello() {
+        assertTrue("JVM" in hello())
+    }
+}
