--- src/jvmTest/kotlin/sample/SampleTestsJVM.kt
@@ -1,11 +1,0 @@
-package sample
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-class SampleTestsJVM {
-  @Test
-  fun testHello() {
-    assertTrue("JVM" in hello())
-  }
-}
