+++ src/linuxMain/kotlin/meta/SampleLinux.kt
@@ -1,0 +1,9 @@
+package meta
+
+actual class Sample {
+    actual fun checkMe() = 7
+}
+
+actual object Platform {
+    actual val name: String = "Native"
+}
