+++ src/linuxTest/kotlin/meta/SampleTestsNative.kt
@@ -1,0 +1,11 @@
+package meta
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class SampleTestsNative {
+    @Test
+    fun testHello() {
+        assertTrue("Native" in hello())
+    }
+}
